Difference between the code on book and web
The offical code we can see on 
https://github.com/Apress/beg-game-dev-w-python-pygame/blob/master/bpwpapcode/Chapter03/events.py
Without line 9 and line 11-12 whitch is about MenoryError, might not affect to this test for running properly.
All in all, this code gives us a simple example showing what kinds of input can be captured:
Keybord and mouse keys up and down, the (x, y) of mouse situation. 
( And my xbox controller too, but only for win button, it might be something missed if we want to fully use our controller)
Still it is excited to see how computer actually doing when you playing games pressing the buttons.